# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ZARIA Builder API
  version: 1.1.0
  description: |
    API Gateway for ZARIA Builder - Text Edition.
    Includes compatibility bridge endpoints for legacy clients.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/exports/{documentId}:
    post:
      tags: [Exports]
      summary: Create export
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        "201":
          description: Export artifact
  /v1/exports/{exportId}:
    get:
      tags: [Exports]
      summary: Get export metadata
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ExportId'
      responses:
        "200":
          description: Export metadata
  /v1/exports/{exportId}/download:
    get:
      tags: [Exports]
      summary: Download export
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ExportId'
      responses:
        "200":
          description: Binary export
...
  /api/modules/{module}/toggle:
    post:
      tags: [Compatibility]
      summary: Legacy module toggle compatibility
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: module
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Toggle result
  /api/memory/search:
    post:
      tags: [Compatibility]
      summary: Legacy memory search compatibility
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
      responses:
        "200":
          description: Search results
  /api/ollama/generate:
    post:
      tags: [Compatibility]
      summary: Legacy transformation compatibility endpoint
      description: Performs structured transformation only; no text generation.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, text]
              properties:
                title:
                  type: string
                text:
                  type: string
                templateId:
                  type: string
                metadata:
                  type: object
                spine:
                  $ref: '#/components/schemas/SpineMetrics'
      responses:
        "200":
          description: Structured transformation payload
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      in: header
      name: x-tenant-id
      required: true
      schema:
        type: string
    DocumentId:
      in: path
      name: documentId
      required: true
      schema:
        type: string
    ExportId:
      in: path
      name: exportId
      required: true
      schema:
        type: string
  schemas:
...
    SpineMetrics:
      type: object
      required: [ad, pm, esi]
      properties:
        ad:
          type: number
          minimum: 0
          maximum: 100
        pm:
          type: number
          minimum: 0
          maximum: 100
        esi:
          type: number
          minimum: 0
          maximum: 100
...
    ExportRequest:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [pdf, epub, docx, bundle]
        includeFormats:
          type: array
          items:
            type: string
            enum: [pdf, epub, docx]
        spine:
          $ref: '#/components/schemas/SpineMetrics'
...